refactor!: move `locate`, `toml`, and `settings` modules into `config` module
